src/requirements.txt
requests  # HTTP library for making API calls.
pandas  # Data manipulation and analysis library.
scikit-learn==1.5.2  # Tools for predictive data analysis.
joblib  # Efficient serialization of Python objects.
sounddevice  # Audio I/O for capturing and playing sound.
vosk  # Offline speech recognition library.
pocketsphinx  # CMU's lightweight speech recognition library.
sentence-transformers  # Library for sentence embeddings and semantic search.
evdev  # Handle input devices such as keyboards, mice (Linux only).
selenium  # Web automation library for controlling browsers.
discord.py  # Discord API wrapper for bot creation.
hyperdb-python  # High-performance database library.
tiktoken 
python-dotenv
pygame
moviepy  # Python library for video editing / playing videos.
sox
#espeak-ng
#alsa-utils
adafruit-pca9685
azure-cognitiveservices-speech
